# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Chess vs Stockfish
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Game {
  id: number;
  pgn: string;
  /** Game result: white, black, or draw */
  result: string;
  /** Player color: white or black */
  playerColor: string;
  /** Stockfish difficulty 1-20 */
  difficulty: number;
  createdAt: string;
}

export interface CreateGameBody {
  pgn: string;
  result: string;
  playerColor: string;
  difficulty: number;
}

export interface ErrorResponse {
  error: string;
}