# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Chess vs Stockfish
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all saved games
 * @summary List saved games
 */
export const ListGamesResponseItem = zod.object({
  id: zod.number(),
  pgn: zod.string(),
  result: zod.string().describe("Game result: white, black, or draw"),
  playerColor: zod.string().describe("Player color: white or black"),
  difficulty: zod.number().describe("Stockfish difficulty 1-20"),
  createdAt: zod.coerce.date(),
});
export const ListGamesResponse = zod.array(ListGamesResponseItem);

/**
 * @summary Save a game
 */
export const CreateGameBody = zod.object({
  pgn: zod.string(),
  result: zod.string(),
  playerColor: zod.string(),
  difficulty: zod.number(),
});

/**
 * @summary Get a game by ID
 */
export const GetGameParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGameResponse = zod.object({
  id: zod.number(),
  pgn: zod.string(),
  result: zod.string().describe("Game result: white, black, or draw"),
  playerColor: zod.string().describe("Player color: white or black"),
  difficulty: zod.number().describe("Stockfish difficulty 1-20"),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a game
 */
export const DeleteGameParams = zod.object({
  id: zod.coerce.number(),
});