# lib/api-zod/src/generated/types/createGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Chess vs Stockfish
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGameBody {
  pgn: string;
  result: string;
  playerColor: string;
  difficulty: number;
}