# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Chess vs Stockfish
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}